Split up setup code based on feedback
